Added a numeric and a range query and all the intermediary services.
git-svn-id: http://ikube.googlecode.com/svn/trunk@1108 89a0fb20-889e-aefa-90f6-7c296b20d5cf

+++ code/core/src/main/java/ikube/index/IndexManager.java
@@ -528,7 +528,7 @@ public final class IndexManager {
 	}
 
 	public static void addNumericField(final String fieldName, final String fieldContent, final Document document, final Store store) {
-		document.add(new NumericField(fieldName, store, true).setLongValue(Long.parseLong(fieldContent)));
+		document.add(new NumericField(fieldName, store, true).setDoubleValue(Double.parseDouble(fieldContent)));
 	}
 
 	public static void addReaderField(final String fieldName, final Document document, final Store store, final TermVector termVector,
